Update video, defaults.
The apple video was previously 120 FPS because the treadmill was too
fast. We always intended to drop the framerate to 60. Update the video
file to be 1/2 the speed at 60 FPS.

Also update the defaults and keepout points to use a 16:9 ratio for both
videos (960x540) for better presentation.

Tested:
manufacturing_demo on x86

Change-Id: I0bc6ca73f166d3275382250b65ec36802e1e42e1