Unify SVG Generation
In the previous pipeline, each branch of the pipeline would use
rsvgoverlay before sending the frame to the mixer. This is inefficent
and also an issue when using the dev board's glsvgoverlay sink.

Creates a new class, SvgGenerator, which lets both appsinks update the
global SVG. It is mutex guarded to ensure both threads don't update the
string before the overlay is updated.

Tested:
manufacturing_demo on x86

Change-Id: I3ebab4a758cfa0b8291793cbfa4f23699937bad2