Indicate safety collisions based on bottom
Currently, any collision between the detected person and the keepout
polygon is reported as a collision in the worker safety demo. This
yields many false positives when a person walks in front of the keepout
region for a high-angle camera.

Add a new flag safety_check_whole_box, to keep this behavior. By
defualt, make this false and only check the bottom of the bounding box
which cooresponds to the worker's feet.

Change-Id: Iba7311690adf906e607bf478bb039fa99f1c8960